+++ .storybook/themeableAttributes.js
@@ -27,6 +27,7 @@ export default {
   'bpk-component-calendar': require('./../packages/bpk-component-calendar').themeAttributes,
   'bpk-component-link': require('./../packages/bpk-component-link').themeAttributes,
   'bpk-component-horizontal-nav': require('./../packages/bpk-component-horizontal-nav').themeAttributes,
+  'bpk-component-modal': require('./../packages/bpk-component-modal').themeAttributes,
   'bpk-component-nudger': require('./../packages/bpk-component-nudger').themeAttributes,
   'bpk-component-spinner': require('./../packages/bpk-component-spinner').themeAttributes,
   'bpk-component-progress': require('./../packages/bpk-component-progress').themeAttributes,