+++ packages/bpk-component-close-button/index.js
@@ -17,5 +17,7 @@
  */
 
 import BpkCloseButton from './src/BpkCloseButton';
+import themeAttributes from './src/themeAttributes';
 
 export default BpkCloseButton;
+export { themeAttributes };