+++ packages/bpk-component-close-button/src/themeAttributes-test.js
@@ -16,8 +16,13 @@
  * limitations under the License.
  */
 
-import BpkModal from './src/BpkModal';
-import themeAttributes from './src/themeAttributes';
+import themeAttributes from './themeAttributes';
 
-export default BpkModal;
-export { themeAttributes };
+describe('themeAttributes', () => {
+  it('exports the expected themeAttributes', () => {
+    expect(themeAttributes).toEqual([
+      'modalCloseButtonColor',
+      'modalCloseButtonHoverColor',
+    ]);
+  });
+});