+++ packages/bpk-docs/src/pages/ModalsPage/ModalsPage.js
@@ -66,10 +66,18 @@ class ModalContainer extends Component {
     return (
       <div>
         <BpkButton onClick={this.onOpen}>{buttonText}</BpkButton>
-        <BpkModal id="my-modal" isOpen={this.state.isOpen} onClose={this.onClose} wide={this.state.hideTitle} {...rest}>
+        <BpkModal
+          id="my-modal"
+          isOpen={this.state.isOpen}
+          onClose={this.onClose}
+          wide={this.state.hideTitle}
+          renderTarget={this.theamableModalTarget}
+          {...rest}
+        >
           <div>{children}</div>
           <BpkButtonLink onClick={this.toggleWidth}>Toggle width</BpkButtonLink>
         </BpkModal>
+        <div ref={(target) => { this.theamableModalTarget = target; }} />
       </div>
     );
   }