+++ packages/bpk-docs/src/themeableAttributes.js
@@ -27,6 +27,7 @@ export default [
   require('bpk-component-calendar').themeAttributes,
   require('bpk-component-link').themeAttributes,
   require('bpk-component-horizontal-nav').themeAttributes,
+  require('bpk-component-modal').themeAttributes,
   require('bpk-component-nudger').themeAttributes,
   require('bpk-component-spinner').themeAttributes,
   require('bpk-component-progress').themeAttributes,